Translate queries into English
For the quiz at re:publica 2023 [1] – it probably won’t use the real
Datenpumpe interface live, but this should still be useful for
preparation.

I also updated the number of results for airports named after women, and
corrected the title of the “number of species” query – it’s actually
counting all species, whether animal plant or other, and IIRC the plants
actually significantly outnumber the animals.

I’ve left the three queries specific to German municipalities as they
are, I don’t think it makes sense to translate them (and they won’t be
featured in the quiz anyways).

[1]: https://re-publica.com/de/session/strange-obscure-and-interesting-facts-wikipedia-quiz