Replace two slow queries with static data
The two queries for “most inherited professions” and “number of women
and Johns in British parliaments” are very slow and can sometimes even
time out. This isn’t really a problem when using the real data pump,
because it doesn’t run queries live during pumping; however, when we’re
thinking of using the datenpumpe website directly at re:publica, it’s
more important that these queries are fast, so replace them with a
static result set temporarily. (This will be reverted later.)